requirements.txt
Django==4.0.6
opentelemetry-api==1.13.0
opentelemetry-distro==0.34b0
opentelemetry-exporter-otlp-proto-http==1.13.0
opentelemetry-instrumentation==0.34b0
opentelemetry-instrumentation-django==0.34b0
opentelemetry-sdk==1.13.0
